requirements.txt
pyyaml
uritemplate
djangorestframework
django
